pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "carryoncarlos"
version = "0.1.0"
dependencies = [
    "google-cloud-logging",
    "flask",
    "flask-cors",
    "openai",
    "python-dotenv"
]

[tool.setuptools.packages.find]
include = ["app"]
